refactor : Notification을 상속 관계로 매핑 (#47)
