+++ breaditnow-domain/src/main/java/com/breaditnow/domain/domain/notification/dto/NotificationDTO.java
@@ -1,0 +1,4 @@
+package com.breaditnow.domain.domain.notification.dto;
+
+public class NotificationDTO {
+}
