+++ breaditnow-domain/src/main/java/com/breaditnow/domain/domain/notification/entity/CustomerAlertNotification.java
@@ -1,34 +1,21 @@
 package com.breaditnow.domain.domain.notification.entity;
 
 import static jakarta.persistence.FetchType.*;
-import static jakarta.persistence.GenerationType.*;
-import static lombok.AccessLevel.*;
 
 import com.breaditnow.domain.domain.customer.entity.Customer;
 import com.breaditnow.domain.domain.product.entity.Product;
-import com.breaditnow.domain.global.entity.BaseEntity;
 
+import jakarta.persistence.DiscriminatorValue;
 import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.Id;
 import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
 import lombok.Builder;
 import lombok.Getter;
-import lombok.NoArgsConstructor;
 
 @Entity
+@DiscriminatorValue("alert")
 @Getter
-@NoArgsConstructor(access = PROTECTED)
-public class CustomerAlertNotification extends BaseEntity {
-	@Id
-	@GeneratedValue(strategy = IDENTITY)
-	private Long id;
-
-	@ManyToOne(fetch = LAZY)
-	@JoinColumn(name = "customer_id", nullable = false)
-	private Customer customer;
-
+public class CustomerAlertNotification extends CustomerNotification {
 	@ManyToOne(fetch = LAZY)
 	@JoinColumn(name = "product_id")
 	private Product product;
@@ -41,27 +28,14 @@ public class CustomerAlertNotification extends BaseEntity {
 
 	private String productName;
 
-	private boolean isRead;
-
-	private boolean isActive = true;
-
 	@Builder
 	public CustomerAlertNotification(Customer customer, Product product, Integer remainingCount, Integer alertCount,
-		String bakeryName, String productName, boolean isRead) {
-		this.customer = customer;
-		this.bakeryName = bakeryName;
+		String bakeryName, String productName) {
+		super(customer, true, false);
 		this.product = product;
-		this.productName = productName;
 		this.remainingCount = remainingCount;
 		this.alertCount = alertCount;
-		this.isRead = isRead;
-	}
-
-	public void changeIsRead(boolean isRead) {
-		this.isRead = isRead;
-	}
-
-	public void changeIsActive(boolean isActive) {
-		this.isActive = isActive;
+		this.bakeryName = bakeryName;
+		this.productName = productName;
 	}
 }
