+++ breaditnow-domain/src/main/java/com/breaditnow/domain/domain/notification/repository/UnifiedNotificationRepository.java
@@ -1,0 +1,7 @@
+package com.breaditnow.domain.domain.notification.repository;
+
+import org.springframework.stereotype.Repository;
+
+@Repository
+public class UnifiedNotificationRepository {
+}
